feat(cubejs-cli): Extract auth command to own module
